Create the dataset for the users
This change creates the dataset for the users.
We can now create the outputs as follows:
`make dataset`

This takes some time to run (approx 30 mins). After this completes,
running:
`make classify`

creates a decision tree for the dataset. The dataset is also saved in
this change so you don't have to create it over and over again. It has
no identifying information.